chore rename instruction to transaction
