+++ governance/program/src/processor/mod.rs
@@ -12,7 +12,7 @@ mod process_create_realm;
 mod process_create_token_governance;
 mod process_create_token_owner_record;
 mod process_deposit_governing_tokens;
-mod process_execute_instruction;
+mod process_execute_transaction;
 mod process_finalize_vote;
 mod process_flag_transaction_error;
 mod process_insert_transaction;
@@ -41,7 +41,7 @@ use process_create_realm::*;
 use process_create_token_governance::*;
 use process_create_token_owner_record::*;
 use process_deposit_governing_tokens::*;
-use process_execute_instruction::*;
+use process_execute_transaction::*;
 use process_finalize_vote::*;
 use process_flag_transaction_error::*;
 use process_insert_transaction::*;
